// src/types/auth/authErrors.ts
export enum AuthErrors {
    EmailAlreadyExists = "EMAIL_ALREADY_EXISTS",
    InvalidPassword = "INVALID_PASSWORD",
    EmailNotFound = "EMAIL_NOT_FOUND",
    InvalidEmail = "INVALID_EMAIL",
    InvalidUid = "INVALID_UID",
    EmailMissing = "EMAIL_MISSING",
    PasswordMissing = "PASSWORD_MISSING",
    UidMissing = "UID_MISSING",
    InvalidLoginCredentials = "INVALID_LOGIN_CREDENTIALS",
    TokenMissing = "TOKEN_MISSING",        
    InvalidToken = "INVALID_TOKEN",
    UnexpectedError = "UNEXPECTED_ERROR"
}